+++ fastapi__template/dependencies/load_env.py
@@ -13,6 +13,6 @@ from dotenv import load_dotenv
 
 def load_env():
     """Load environment variables."""
-    env_path = os.path.join(os.path.dirname(__file__), "env_files/.env")
+    env_path = os.path.join(os.path.dirname(__file__), "../.env")
     env_path = os.path.abspath(env_path)
     load_dotenv(env_path)
